# Implement date-level caching for JSON decoding