--- a/Tests/StreamChatTests/Utils/JSONDecoder_Tests.swift
+++ b/Tests/StreamChatTests/Utils/JSONDecoder_Tests.swift
@@ -7,20 +7,8 @@ import Foundation
 @testable import StreamChatTestTools
 import XCTest
 
-final class JSONDecoderTests: XCTestCase {
-    private var decoder: JSONDecoder!
-
-    private let key = "date"
-
-    override func setUp() {
-        decoder = .default
-        super.setUp()
-    }
-
-    override func tearDown() {
-        decoder = nil
-        super.tearDown()
-    }
+final class JSONDecoder_Tests: XCTestCase {
+    private var decoder: JSONDecoder = .default
 
     func test_throwsException_whenDecodingDateFromEmptyString() {
         checkDecodingDateThrowException(dateString: "")
@@ -34,8 +22,8 @@ final class JSONDecoderTests: XCTestCase {
         checkDecodingDateThrowException(dateString: "2020-09-30T19:51:17")
     }
 
-    func test_decodes_whenDecodingDateFromRFC3339DateWithMilliseconds() {
-        checkDateIsDecodingToComponents(
+    func test_decodes_whenDecodingDateFromRFC3339DateWithMilliseconds() throws {
+        try checkDateIsDecodingToComponents(
             dateString: "2020-08-24T17:28:04.123Z",
             year: 2020,
             month: 8,
@@ -47,8 +35,8 @@ final class JSONDecoderTests: XCTestCase {
         )
     }
 
-    func test_decodes_whenDecodingDateFromRFC3339DateWithEmptyMilliseconds() {
-        checkDateIsDecodingToComponents(
+    func test_decodes_whenDecodingDateFromRFC3339DateWithEmptyMilliseconds() throws {
+        try checkDateIsDecodingToComponents(
             dateString: "2002-12-02T15:11:12Z",
             year: 2002,
             month: 12,
@@ -59,8 +47,8 @@ final class JSONDecoderTests: XCTestCase {
         )
     }
 
-    func test_decodes_whenDecodingDateFromRFC3339DateWithMinusTimezone() {
-        checkDateIsDecodingToComponents(
+    func test_decodes_whenDecodingDateFromRFC3339DateWithMinusTimezone() throws {
+        try checkDateIsDecodingToComponents(
             dateString: "2002-10-02T07:12:13-03:00",
             year: 2002,
             month: 10,
@@ -71,8 +59,8 @@ final class JSONDecoderTests: XCTestCase {
         )
     }
 
-    func test_decodes_whenDecodingDateFromRFC3339DateWithPlusTimezone() {
-        checkDateIsDecodingToComponents(
+    func test_decodes_whenDecodingDateFromRFC3339DateWithPlusTimezone() throws {
+        try checkDateIsDecodingToComponents(
             dateString: "2002-10-02T10:12:13+02:00",
             year: 2002,
             month: 10,
@@ -83,8 +71,8 @@ final class JSONDecoderTests: XCTestCase {
         )
     }
 
-    func test_decodes_whenDecodingDateFromRFC3339DateWithPlusZeroTimezone() {
-        checkDateIsDecodingToComponents(
+    func test_decodes_whenDecodingDateFromRFC3339DateWithPlusZeroTimezone() throws {
+        try checkDateIsDecodingToComponents(
             dateString: "2002-10-02T10:12:13+00:00",
             year: 2002,
             month: 10,
@@ -95,8 +83,8 @@ final class JSONDecoderTests: XCTestCase {
         )
     }
 
-    func test_decodes_whenDecodingDateFromRFC3339DateWithMinusZeroTimezone() {
-        checkDateIsDecodingToComponents(
+    func test_decodes_whenDecodingDateFromRFC3339DateWithMinusZeroTimezone() throws {
+        try checkDateIsDecodingToComponents(
             dateString: "2002-10-02T10:12:13-00:00",
             year: 2002,
             month: 10,
@@ -107,8 +95,8 @@ final class JSONDecoderTests: XCTestCase {
         )
     }
 
-    func test_decodes_whenDecodingDateBefore1970() {
-        checkDateIsDecodingToComponents(
+    func test_decodes_whenDecodingDateBefore1970() throws {
+        try checkDateIsDecodingToComponents(
             dateString: "1936-10-02T10:12:13Z",
             year: 1936,
             month: 10,
@@ -118,13 +106,83 @@ final class JSONDecoderTests: XCTestCase {
             second: 13
         )
     }
-}
+    
+    func test_defaultDecoder_isStreamDecoder() {
+        // Assert that default decoder we use is the stream decoder
+        XCTAssert(JSONDecoder.default === JSONDecoder.stream)
+        XCTAssert(type(of: JSONDecoder.stream) == StreamJSONDecoder.self)
+        
+        // Assert the default parameters are correctly initialized
+        XCTAssertEqual(JSONDecoder.stream.dateCache.countLimit, 5000)
+        XCTAssertEqual(JSONDecoder.stream.iso8601formatter.formatOptions, [.withFractionalSeconds, .withInternetDateTime])
+    }
+    
+    func test_datesAreCached() throws {
+        final class ISO8601DateFormatter_Spy: ISO8601DateFormatter {
+            var dateFromStringCalledCounter: Int = 0
+            
+            override func date(from string: String) -> Date? {
+                dateFromStringCalledCounter += 1
+                return super.date(from: string)
+            }
+        }
+        
+        final class NSCache_Spy: NSCache<NSString, NSDate> {
+            var setObjectCalledCounter: Int = 0
+            var getObjectCalledCounter: Int = 0
+            
+            override func object(forKey key: NSString) -> NSDate? {
+                getObjectCalledCounter += 1
+                return super.object(forKey: key)
+            }
+            
+            override func setObject(_ obj: NSDate, forKey key: NSString) {
+                setObjectCalledCounter += 1
+                super.setObject(obj, forKey: key)
+            }
+        }
+        
+        // Given a decoder with spy dateFormatter and cache
+        let dateFormatter = ISO8601DateFormatter_Spy()
+        dateFormatter.formatOptions = [.withFractionalSeconds, .withInternetDateTime]
+        let dateCache = NSCache_Spy()
+        let decoder = StreamJSONDecoder(dateFormatter: dateFormatter, dateCache: dateCache)
+        
+        // When we decode a payload with repeated dates
+        let repeatedDate = "2020-06-09T08:10:40.800912Z" // If you change this, make sure to change `actualDecodedDate` below
+        let jsonPayload = """
+            {
+                "date1": "\(repeatedDate)",
+                "date2": "\(repeatedDate)",
+                "date3": "\(repeatedDate)",
+                "date4": "\(repeatedDate)",
+                "date5": "\(repeatedDate)"
+            }
+        """
+        let dateDict = try decoder.decode([String: Date].self, from: jsonPayload.data(using: .utf8)!)
+        
+        // Then we should only decode the date once and use the cache
+        XCTAssertEqual(dateFormatter.dateFromStringCalledCounter, 1)
+        XCTAssertEqual(dateCache.setObjectCalledCounter, 1)
+        XCTAssertEqual(dateCache.getObjectCalledCounter, 5)
+        
+        // The actual decoded date must match the date JSONDecoder decoded and cached
+        let actualDecodedDate = Date(timeIntervalSince1970: 1_591_690_240.8)
+        XCTAssertEqual(dateCache.object(forKey: repeatedDate as NSString)?.timeIntervalSince1970, actualDecodedDate.timeIntervalSince1970)
+        
+        // All dates must be decoded
+        XCTAssertEqual(dateDict.keys.count, 5)
+        for (_, value) in dateDict {
+            XCTAssertEqual(value, actualDecodedDate)
+        }
+    }
 
-// MARK: Helpers
+    // MARK: Helpers
 
-extension JSONDecoderTests {
+    private let dateKey = "date"
+    
     private func json(dateString: String) -> String {
-        "{\"\(key)\":\"\(dateString)\"}"
+        "{\"\(dateKey)\":\"\(dateString)\"}"
     }
 
     private func checkDateIsDecodingToComponents(
@@ -135,17 +193,19 @@ extension JSONDecoderTests {
         hour: Int,
         minute: Int,
         second: Int,
-        fractionalSeconds: Int? = nil
-    ) {
+        fractionalSeconds: Int? = nil,
+        file: StaticString = #filePath,
+        line: UInt = #line
+    ) throws {
         // Given
         let dateJson = json(dateString: dateString)
         let data = dateJson.data(using: .utf8)!
 
         // When
-        let decoded: [String: Date] = try! decoder.decode([String: Date].self, from: data)
+        let decoded: [String: Date] = try decoder.decode([String: Date].self, from: data)
 
         // Then
-        let decodedDate = decoded[key]!
+        let decodedDate = decoded[dateKey]!
 
         // Use GMT calendar, to test on GMT+0 timezone
         let components = Calendar.gmtCalendar.dateComponents(
@@ -153,12 +213,12 @@ extension JSONDecoderTests {
             from: decodedDate
         )
 
-        XCTAssertEqual(components.year, year)
-        XCTAssertEqual(components.month, month)
-        XCTAssertEqual(components.day, day)
-        XCTAssertEqual(components.hour, hour)
-        XCTAssertEqual(components.minute, minute)
-        XCTAssertEqual(components.second, second)
+        XCTAssertEqual(components.year, year, file: file, line: line)
+        XCTAssertEqual(components.month, month, file: file, line: line)
+        XCTAssertEqual(components.day, day, file: file, line: line)
+        XCTAssertEqual(components.hour, hour, file: file, line: line)
+        XCTAssertEqual(components.minute, minute, file: file, line: line)
+        XCTAssertEqual(components.second, second, file: file, line: line)
 
         if let fractional = fractionalSeconds {
             let nanosecondsInMillisecond = 1_000_000
@@ -171,11 +231,11 @@ extension JSONDecoderTests {
                 fractionalResult += 1
             }
 
-            XCTAssertEqual(fractionalResult, fractional)
+            XCTAssertEqual(fractionalResult, fractional, file: file, line: line)
         }
     }
 
-    private func checkDecodingDateThrowException(dateString: String) {
+    private func checkDecodingDateThrowException(dateString: String, file: StaticString = #filePath, line: UInt = #line) {
         // Given
         let dateJson = json(dateString: "")
         let data = dateJson.data(using: .utf8)!
@@ -185,7 +245,7 @@ extension JSONDecoderTests {
             _ = try decoder.decode([String: Date].self, from: data)
         } catch {
             // Then
-            XCTAssertNotNil(error)
+            XCTAssertNotNil(error, file: file, line: line)
         }
     }
 }