@b-onc 2022-06-28
@polqf Done! Decoder is much more testable now